Adjusting namespaces, moving rendering logic out of engine namespace
