+++ SharpPhysics.Demo/DemoProps/SandboxDebugUI.cs
@@ -1,11 +1,11 @@
 #nullable enable
 using System.Numerics;
-using physics.Engine;
 using physics.Engine.Core;
-using physics.Engine.Rendering;
-using physics.Engine.Rendering.UI;
-using physics.Engine.Shaders;
 using SFML.Graphics;
+using SharpPhysics.Engine.Core;
+using SharpPhysics.Rendering;
+using SharpPhysics.Rendering.Shaders;
+using SharpPhysics.Rendering.UI;
 
 namespace SharpPhysics.Demo.DemoProps;
