+++ SharpPhysics.Demo/PlatformerGame.cs
@@ -1,15 +1,15 @@
 #nullable enable
 using System.Numerics;
-using physics.Engine;
 using physics.Engine.Classes.ObjectTemplates;
 using physics.Engine.Core;
 using physics.Engine.Input;
 using physics.Engine.Objects;
-using physics.Engine.Rendering;
-using physics.Engine.Shaders;
 using SharpPhysics.Engine.Player;
 using SFML.Graphics;
 using SFML.Window;
+using SharpPhysics.Engine.Core;
+using SharpPhysics.Rendering.Shaders;
+using SharpPhysics.Rendering;
 
 namespace SharpPhysics.Demo;
