+++ SharpPhysics.Demo/SettingsGame.cs
@@ -3,11 +3,11 @@ using System.Numerics;
 using physics.Engine;
 using physics.Engine.Core;
 using physics.Engine.Input;
-using physics.Engine.Rendering;
-using physics.Engine.Rendering.UI;
 using SFML.Graphics;
 using SFML.Window;
 using SharpPhysics.Demo.Settings;
+using SharpPhysics.Rendering;
+using SharpPhysics.Rendering.UI;
 
 namespace SharpPhysics.Demo;
