+++ SharpPhysics.Tests/CollisionTests.cs
@@ -1,11 +1,11 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Numerics;
-using physics.Engine;
 using physics.Engine.Shapes;
 using physics.Engine.Structs;
 using System;
 using physics.Engine.Classes;
 using physics.Engine.Objects;
+using SharpPhysics.Engine.Core;
 
 namespace SharpPhysics.Tests
 {
