+++ physics/Rendering/Shaders/SFMLPolyRainbowShader.cs
@@ -4,9 +4,8 @@ using physics.Engine.Objects;
 using physics.Engine.Helpers;
 using SFML.System;
 using System;
-using physics.Engine.Shaders.Helpers;
 
-namespace physics.Engine.Shaders
+namespace SharpPhysics.Rendering.Shaders
 {
     public class SFMLPolyRainbowShader : SFMLShader
     {
