+++ physics/Rendering/UI/UiSlider.cs
@@ -3,7 +3,7 @@ using SFML.Graphics;
 using System;
 using System.Numerics;
 
-namespace physics.Engine.Rendering.UI
+namespace SharpPhysics.Rendering.UI
 {
     public class UiSlider : UiElement, IUiClickable, IUiDraggable
     {
